sorting nearly done

+++ Students/BURSKENS/TP1/include/sorting_algorithms.hpp
@@ -7,4 +7,4 @@ void bubble_sort(int* array, int length);
 void insertion_sort(int* array, int length);
 
 
-void merge_sort(int* array, int length);
+void merge_sort(int* array, size_t length);
